fucking finally
bullets now shoot beautifully. needed to make a central position due to the player object moving. when  it moved, its fire position tried to accommodate. now that it shoots from position {0,0} and the bullet draw position is from player object, it shoots from the player. this may cause collision issues in the future, but im gonna take my win for now