collision fixed
use #defines to center the rectangle on the size of each entity. bullet cannot be attached to collision component due to the for loop. add width and height. keep collision component for other entities.

also figure out how to replace int i for bullet so it can use the fucking transform component